Add files via upload
Some code snippets from Tides of Destiny, a group project made as a student project during education. 
You can see the current game on our itch.io: 
https://tod-pirates.itch.io/tides-of-destiny
(Warning: this was made by a group of beginners with a deadline, and is still very much a work in progress with many bugs and issues to be solved)

Not all of the code in the project is mine, this repository mostly consists of parts I have personally worked on in some capacity. This includes the Equipment and Items system / Player Stats, Shop functionality, and some help and fixes for Enemy and Player behaviors and UI scripts.