Survivors Mechanics
Some code for a Vampire Survivors imitation game I began as a practice project some months ago.
I remade some of these systems in my other RogueLite project, and these are far from finished. But this project was based on an idea of combining different weapon parts to modify your build providing a high level of customization. However, due to deadlines on schoolwork and other responsibilities it remains unfinished for now